collections exercise

+++ src/main/java/grades/Student.java
@@ -18,14 +18,14 @@ public class Student {
     }
 
     // adds the given grade to the grades property
-    public void addGrade(int grade) {
+    public void addGrade(Integer grade) {
         this.grades.add(grade);
     }
 
     // returns the average of the students grades
     public double getGradeAverage() {
-        int sum = 0 ;
-        for (Integer grade : this.grades) {
+        double sum = 0 ;
+        for (double grade : this.grades) {
             sum += grade;
         }
         return sum / this.grades.size();
